+++ PerformanceCalculatorGUI/Components/TextBoxes/LimitedLabelledFractionalNumberBox.cs
@@ -1,16 +1,21 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Globalization;
 using osu.Framework.Bindables;
+using osu.Framework.Extensions;
 using osu.Game.Graphics.UserInterface;
-using osu.Game.Graphics.UserInterfaceV2;
 
-namespace PerformanceCalculatorGUI.Components
+namespace PerformanceCalculatorGUI.Components.TextBoxes
 {
-    internal class LimitedLabelledNumberBox : LabelledNumberBox
+    internal class LimitedLabelledFractionalNumberBox : ExtendedLabelledTextBox
     {
-        internal class LimitedNumberBox : OsuNumberBox
+        internal class FractionalNumberBox : OsuTextBox
         {
+            protected override bool AllowIme => false;
+
+            protected override bool CanAddCharacter(char character) => character.IsAsciiDigit() || character == CultureInfo.CurrentCulture.NumberFormat.NumberDecimalSeparator[0];
+
             protected override void OnUserTextAdded(string added)
             {
                 base.OnUserTextAdded(added);
@@ -22,9 +27,9 @@ namespace PerformanceCalculatorGUI.Components
                     textToParse = PlaceholderText.ToString();
                 }
 
-                if (int.TryParse(textToParse, out int parsed))
+                if (double.TryParse(textToParse, out double parsed))
                 {
-                    if (parsed >= (MinValue ?? int.MinValue) && parsed <= (MaxValue ?? int.MaxValue))
+                    if (parsed >= MinValue && parsed <= MaxValue)
                     {
                         Value.Value = parsed;
                         return;
@@ -44,7 +49,7 @@ namespace PerformanceCalculatorGUI.Components
                     textToParse = PlaceholderText.ToString();
                 }
 
-                if (int.TryParse(textToParse, out int parsed))
+                if (double.TryParse(textToParse, out double parsed))
                 {
                     Value.Value = parsed;
                     return;
@@ -53,36 +58,25 @@ namespace PerformanceCalculatorGUI.Components
                 Value.Value = default;
             }
 
-            public int? MaxValue { get; set; }
+            public double MaxValue { get; set; }
 
-            public int? MinValue { get; set; }
+            public double MinValue { get; set; }
 
-            public Bindable<int> Value { get; } = new Bindable<int>();
+            public Bindable<double> Value { get; } = new Bindable<double>();
         }
 
-        protected override OsuTextBox CreateTextBox() => new LimitedNumberBox();
+        protected override OsuTextBox CreateTextBox() => new FractionalNumberBox();
 
-        public int? MaxValue
+        public double MaxValue
         {
-            set => ((LimitedNumberBox)Component).MaxValue = value;
+            set => ((FractionalNumberBox)Component).MaxValue = value;
         }
 
-        public int? MinValue
+        public double MinValue
         {
-            set => ((LimitedNumberBox)Component).MinValue = value;
+            set => ((FractionalNumberBox)Component).MinValue = value;
         }
 
-        public Bindable<int> Value => ((LimitedNumberBox)Component).Value;
-
-        public bool CommitOnFocusLoss
-        {
-            get => Component.CommitOnFocusLost;
-            set => Component.CommitOnFocusLost = value;
-        }
-
-        public LimitedLabelledNumberBox()
-        {
-            CornerRadius = ExtendedLabelledTextBox.CORNER_RADIUS;
-        }
+        public Bindable<double> Value => ((FractionalNumberBox)Component).Value;
     }
 }
