+++ PerformanceCalculatorGUI/PerformanceCalculatorGUI.csproj
@@ -6,10 +6,10 @@
     <LangVersion>latest</LangVersion>
   </PropertyGroup>
   <ItemGroup>
-	  <PackageReference Include="ppy.osu.Game" Version="2022.719.0" />
-	  <PackageReference Include="ppy.osu.Game.Rulesets.Osu" Version="2022.719.0" />
-	  <PackageReference Include="ppy.osu.Game.Rulesets.Taiko" Version="2022.719.0" />
-	  <PackageReference Include="ppy.osu.Game.Rulesets.Catch" Version="2022.719.0" />
-	  <PackageReference Include="ppy.osu.Game.Rulesets.Mania" Version="2022.719.0" />
+    <ProjectReference Include="..\osu\osu.Game.Rulesets.Catch\osu.Game.Rulesets.Catch.csproj" />
+    <ProjectReference Include="..\osu\osu.Game.Rulesets.Mania\osu.Game.Rulesets.Mania.csproj" />
+    <ProjectReference Include="..\osu\osu.Game.Rulesets.Osu\osu.Game.Rulesets.Osu.csproj" />
+    <ProjectReference Include="..\osu\osu.Game.Rulesets.Taiko\osu.Game.Rulesets.Taiko.csproj" />
+    <ProjectReference Include="..\osu\osu.Game\osu.Game.csproj" />
   </ItemGroup>
 </Project>
