+++ PerformanceCalculatorGUI/Screens/BeatmapLeaderboardScreen.cs
@@ -26,6 +26,7 @@ using osu.Game.Scoring;
 using osu.Game.Users;
 using osu.Game.Utils;
 using PerformanceCalculatorGUI.Components;
+using PerformanceCalculatorGUI.Components.TextBoxes;
 using PerformanceCalculatorGUI.Configuration;
 
 namespace PerformanceCalculatorGUI.Screens
